+++ cmd/internal/shared/devicetemplateconverter/config.go
@@ -19,4 +19,6 @@ import (
 )
 
 // DefaultDeviceTemplateConverterConfig is the default configuration for the Device Template Converter.
-var DefaultDeviceTemplateConverterConfig = devicetemplateconverter.Config{}
+var DefaultDeviceTemplateConverterConfig = devicetemplateconverter.Config{
+	Enabled: []string{"the-things-stack"},
+}